Added a quick fix to get the runCommandAndLogOutput method to work.
For some reason the forked child process was not detecting the directory
path for the shell script which it needed to execute. For now appended
the full directory path (using the workdir variable - No hardcoding) to
the filename (temp-script.sh).

Would revisit and see what the proper fix would be.