Multiple changes to the jsp files to handle path issues.
These jsp files had the path to static resources (.js, images etc)
not working after the migration. This maybe because we are directly
setting the tomcat document root to the exploded documents folder
rather than the war file. Needed to put this temporary chnage to get
things working. May revert if not needed once we go back to the war
file mode.