Corrected compare() definitions
Previous compare() declarations violated the general contract; they did not
return 0 when a == b. This is checked in Java7+ (I think). Not sure why it's
*not* breaking things right now, because it should raise an
IllegalArgumentException "Comparison method violates its general contract!".
But considering that this problem already broke the kNN+ Models page in the
past, it's best to take care of this now.