Disabled mtry, nodesize parameters for RF
The parameters "mtry" and "nodesize" correspond to "Descriptors Sampled per
Node" and "Minimum Terminal Node Size", respectively. Currently, these are fixed
values that do not differ based on classification / regression and on the number
of descriptors, both of which negatively affect the outcome of RF. As a result,
it has been judged that the defaults work better, and so these options have been
removed from the modeling form and from related code for now.

In the future, these options should be reinstated under an "advanced options"
section, because they should only be used iff the user completely understands
what they are getting into.