jQuery to replace Dojo for ViewDataset
Begone, deprecated libraries of yore! Replacing the Struts2-Dojo tags with plain
ol' jQuery because struts2-dojo-plugin has been deprecated for some time. The
old version of dojo.js was causing issues with tabbed views in IE10; a patch to
one file fixed most of it, but it didn't work for this page, so it was
rewritten with jQuery.

Eventually this will be done for the other tabbed views on the website.