Hotfix for randomForest prediction issue
The last two lines of an X file are either the min/max values or the avg/stddev
values depending on the scaling type used. Problem is that our randomForest
prediction script thinks that these are compounds and crashes on them.

Due to time constraints it's not possible to amend the R code so that
1) prediction works, while
2) not breaking RF modeling
So this info has been left out of the X file for now.