Implemented SVM SMILES prediction
Somewhat rough; some duplicate code from Svm.readPredictionOutput(), but due to
the way that PredictSmilesSDF() is coded the standard deviation/mean
calculations are already done in PredictSmilesSDF(), and so
Svm.readPredictionOutput() wasn't used.

In the future we would want to separate the logic of calculating predictions
into separate classes that specify what to do for a particular modeling method.