Deleted bad code for predictor deletion after making prediction
Doesn't work for public datasets or public predictors, nor for uploaded
descriptors, so what's the point? Pending an executive decision on whether or
not to keep any part of the directory (if there's something worth keeping it's
the .preds file, which contains the actual prediction output). Predictions are
stored in the database, so the files are just there for debugging purposes.