Made "error summary" always visible
Before this message line was hidden unless the descriptor generation failed;
however, it occasionally shows useful warnings (e.g. CDK, where it shows
specific compounds for which descriptor generation failed). So it's more useful
to just display it in all cases.

Later there should be an attempt to hide the line if the corresponding property
is empty or null.