Fixed polling problem
The result of a query appears to get cached if the session is reused, meaning
that the value of jobCompleted will never be seen as "YES", even if it has
changed in the database. The fix creates and destroys a Session object on each
poll, which is hardly ideal, but returns the correct result.

Consequently, polling rate was upped from 100 ms to 1 second in a past commit
to avoid excessive creation/destruction of sessions.