Implemented remainder of generatePredictor method
Finally the NYI comes off! Predictor job creation is currently lying free in the
main method, but it'll be refactored out like the Dataset job creation soon.

Note that this method doesn't care about whether or not the Predictor is
completely finished executing when it returns. Might need to amend these
semantics later if the prediction endpoint needs it.