Send all LSF jobs to 'patrons' queue
Before we would check the number of jobs in 'patrons' and then decide whether to
send new LSF jobs to 'patrons' or to 'idle'. Since jobs are suspended or
preempted for very long periods of time in 'idle', it is better to submit jobs
to 'patrons' and let them stay in PEND rather than send any jobs to the 'idle'
queue.

Removed LsfOperations as it only contained a function that checked the number of
jobs in the 'patrons' queue, which is no longer needed.

Removed the value Constants.PATRONSQUEUESLOTS as it is no longer checked.