Fixed NullPointerException when predictor doesn't exist
When attempting to delete a prediction, if the predictor that created the
prediction no longer exists in the database (due to a forced delete in the MySQL
console or due to an inconsistency), a NullPointerException is thrown.

With the change, a warning is given, but the process will continue without
attempting to delete the expected predictor.