Fixed bad comparison in n-fold ext validation
It's never null since the ArrayList is instantiated. Should be checking for no
values returned. (This happens if no models were generated that passed cutoffs
for a fold.)

Since the check failed, it would go on to return NaN (from 0/0) which would
propagate through to externalPredictionAccuracyAvg, which reported "NaN +/- NaN"
for the parent predictor.