Added validity check for KNN+ models.tbl file
Sometimes KNN+ repeats data on an extra, spurious line which is interpreted by
the KNN+ prediction code as another model. Before we predict the external set we
should check the models.tbl for these spurious lines and remove them before they
can propagate to the database and to KNN+'s prediction code.