Replaced MarvinSketch applet with JSME
The various *.cache.{html,png} items are to assist loading times by having
pre-cached versions of applets and resources available.

This removes one of the two places that we use Java (the other being the 3d
molecule viewer). Load times should be much faster for the Make Predictions page
as well.