Split up CSS files and minify them later
This uses maven-minify-plugin to join together the now-broken-up CSS files. The
chembench.css mega-file was getting too unwieldy so this had to happen sooner
or later.

I ran into some problems dealing with minified JS so that's been taken out for
now, but will hopefully be implemented in the future.