Fix vertical clipping issues with headers
This was initially meant to truncate excessively long object names for the
detail pages so that the button wouldn't overflow onto the next line, but it's
also clipping glyphs with descenders (e.g. "g" on the home page).

Ideally we could specify something like
    overflow-x: hidden;
    overflow-y: visible;
    ...
But that isn't working out, at least in latest Chrome.