Refresh all tables, not just jobs
A common usage pattern is to wait until a job is finished, then click the
appropriate object-type tab to go find it. But if non-job tables aren't updated
then the user won't be able to find the newly completed job.

To that end we should periodically update all the tables, not just the job
queues.