Fix issue with flashing no-objects message
The initial draw needs to happen and THEN the callback needs to be registered,
otherwise you see the "no objects..." message for a split second which is
distracting.

Also this is more easily maintained, as it's probably not a good idea to be
using drawCallback in case it's ever modified by someone else (as the callback
is then lost).