Show CCR avg/stddev properly for nfold predictors
currentFoldNumber's getter had a strange method signature (returned Integer
though the variable was declared String).

Display bug was because currentFoldNumber was treated like a String in this
instance despite an Integer being returned. It made more sense to go ahead and
change currentFoldNumber's type over to Integer. Less chance for this mistake
to be made again.