Adapt to support ChemDoodle instead of Marvin
So MarvinView relies on Java applets, which due to Chrome's deprecating NPAPI
no longer work on Chrome (our audience's most used browser) and applets are
generally a pain to deal with anyway. ChemDoodle works via WebGL and js which
is a lot easier to deal with and, most importantly, supports Chrome.