Convert structures to 3d in dataset creation
convert2Dto3D actually takes quite a while to complete for a single compound,
so trying to convert on the fly whenever someone clicks a compound in the list
view isn't really practical. We can leave it for past datasets, but for future
datasets we ought to do this ahead of time.